feat: add Places page with dynamic category selection and modal details
- Created Places.vue component for displaying a list of places with category filtering.
- Implemented places.js for managing data and methods related to places.
- Added DialogPlaces.vue component for showing detailed information about selected places in a modal.
- Enhanced user experience with responsive design and improved interaction through category selection and modal dialogs.
